+++ eyes_selenium/applitools/selenium/eyes.py
@@ -84,6 +84,7 @@ class Eyes(object):
         if self.configure.is_disabled:
             pass
         else:
+            self._runner._set_connection_config(self.configure)  # noqa, friend
             self._driver = driver
             self._eyes_ref = self._commands.manager_open_eyes(
                 self._runner._ref,  # noqa
@@ -466,11 +467,11 @@ class Eyes(object):
         self._driver = None
         if wait_result:
             results = demarshal_test_results(results, self.configure)
-            for r in results:
-                log_session_results_and_raise_exception(raise_ex, r)
-            return results[0]  # Original interface returns just one result
-        else:
-            return None
+            if results:  # eyes are already aborted by closed runner
+                for r in results:
+                    log_session_results_and_raise_exception(raise_ex, r)
+                return results[0]  # Original interface returns just one result
+        return None
 
     def _abort(self, wait_result):
         # type: (bool) -> Optional[List[TestResults]]
@@ -481,9 +482,12 @@ class Eyes(object):
             self._eyes_ref = None
             self._driver = None
             if wait_result:
-                return demarshal_test_results(results, self.configure)
-            else:
-                return None
+                results = demarshal_test_results(results, self.configure)
+                if results:  # abort after close does not return results
+                    for r in results:
+                        log_session_results_and_raise_exception(False, r)
+                    return results[0]  # Original interface returns just one result
+            return None
 
     def __getattr__(self, item):
         return getattr(self.configure, item)